working finally haha

+++ vdifile.h
@@ -9,7 +9,7 @@ ssize_t VDIRead(int fd,void* buf,size_t count);
 ssize_t VDIWrite(int fd,void* buf,size_t count);
 off_t VDISeek(int fd,off_t offset,int anchor);
 int fileSize;
-int cursor;
+off_t cursor;
 int file;
 int fd;
 };
